utf8n_to_uvuni(): Move checking for >32 bit code points
This just moves the code that does the checking if the code point is
non-portable to 32-bit machines to later.  These code points aren't
representable at all on EBCDIC platforms.

This will have the effect that on platforms where these aren't
representable, the error returned will be overflow instead of these,
but the moving fixes the problem where the first byte is garbled, and
the input really isn't such a large code point.  Prior to this patch,
the first byte is treated as gospel, and the intervening code points
aren't examined, leaving the pointer to the next input byte incorrectly
advanced too far.