utf8.c: refactor utf8n_to_uvuni()
The prior version had a number of issues, some of which have been taken
care of in previous commits; and some are left to do.

The goal when presented with malformed input is to consume as few bytes
as possible so as to position the input for the next try to the first
possible byte that could be the beginning of a character.  We don't want
to consume too few bytes, so that the next call has us thinking that
what is the middle of a character is really the beginning; nor do we
want to consume too many, so as to skip valid input characters.
The previous code could do both of these in various circumstances.

In some cases it believed that the first byte in a character is correct,
and skipped looking at the rest of the bytes in the sequence.  This is
wrong when just that first byte is garbled.  We have to look at all
bytes in the expected sequence to make sure it hasn't been prematurely
terminated.

Likewise when we get an overflow: we have to keep looking at each byte
in the alleged sequence.  It may be that the initial byte was garbled to
give us an apparent large number, but the actual sequence is shorter
than expected, and there really wouldn't have been an overflow.  We
want to position the pointer for the next call to be the beginning of
the next potentially good character.

This fixes a long-standing TODO from an externally supplied utf8 decode
test suite.

Another bug is that the code was careless about what happens when an
allowed malformation happens. For example, a sequence should not start
with a continuation byte.  If that malformation is allowed, the code
pretends it is a start byte and extracts the length of the sequence from
that.  But pretending it is a start byte is not the same thing as it
being a start byte, and that extracted length is bogus.